+++ include/arch/arc/v2/error.h
@@ -14,6 +14,7 @@
 #ifndef _ARCH_ARC_V2_ERROR_H_
 #define _ARCH_ARC_V2_ERROR_H_
 
+#include <arch/arc/syscall.h>
 #include <arch/arc/v2/exc.h>
 
 #ifdef __cplusplus
@@ -32,10 +33,6 @@ extern void _SysFatalErrorHandler(unsigned int cause, const NANO_ESF *esf);
 #define _NANO_ERR_KERNEL_OOPS (4)       /* Kernel oops (fatal to thread) */
 #define _NANO_ERR_KERNEL_PANIC (5)	/* Kernel panic (fatal to system) */
 
-
-#define _TRAP_S_SCALL_IRQ_OFFLOAD		1
-#define _TRAP_S_CALL_RUNTIME_EXCEPT		2
-#define _TRAP_S_CALL_SYSTEM_CALL		3
 
 /*
  * the exception caused by kernel will be handled in interrupt context
