+++ include/syscall.h
@@ -12,7 +12,7 @@
 #include <zephyr/types.h>
 #include <syscall_list.h>
 #include <syscall_macros.h>
-#include <arch/cpu.h>
+#include <arch/syscall.h>
 
 #ifdef __cplusplus
 extern "C" {
